+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Thank you!
 
 - Milkweed now plays a sound when rustled
 - Milkweed can now be sheared with a Dispenser
+- Glory of the Snow can now be sheared with a Dispenser
 - Fixed Mesoglea liquid rendering extending 16 tiles out instead of only encompassing 1 tile
 - Fixed Jellyfish no longer spawning in Jellyfish Caves
 - Slightly optimized Jellyfish rendering with the rainbow easter egg